+++ core/test/unit/src/config/project.ts
@@ -29,6 +29,7 @@ import { resolve, join } from "path"
 import stripAnsi from "strip-ansi"
 import { keyBy } from "lodash"
 
+const enterpriseDomain = "https://garden.mydomain.com"
 const commandInfo = { name: "test", args: {}, opts: {} }
 
 describe("resolveProjectConfig", () => {
@@ -54,6 +55,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -93,6 +96,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -148,6 +153,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: { foo: "banana" },
         commandInfo,
       })
@@ -223,6 +230,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -287,6 +296,8 @@ describe("resolveProjectConfig", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -316,6 +327,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -350,6 +363,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -403,6 +418,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -479,6 +496,8 @@ describe("resolveProjectConfig", () => {
         artifactsPath: "/tmp",
         branch: "main",
         username: "some-user",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -550,6 +569,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),
@@ -577,6 +598,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         branch: "main",
         username,
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -621,6 +644,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         branch: "main",
         username,
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -664,6 +689,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         branch: "main",
         username,
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -722,6 +749,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -776,6 +805,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -824,6 +855,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -873,6 +906,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -922,6 +957,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -982,6 +1019,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1043,6 +1082,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1076,6 +1117,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: { foo: "banana" },
       commandInfo,
     })
@@ -1108,6 +1151,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1139,6 +1184,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1165,6 +1212,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1225,6 +1274,8 @@ describe("pickEnvironment", () => {
       artifactsPath,
       branch: "main",
       username,
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo,
     })
@@ -1267,6 +1318,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),
@@ -1305,6 +1358,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),
@@ -1340,6 +1395,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),
@@ -1367,6 +1424,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         branch: "main",
         username,
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -1399,6 +1458,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         branch: "main",
         username,
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -1431,6 +1492,8 @@ describe("pickEnvironment", () => {
         artifactsPath,
         username,
         branch: "main",
+        loggedIn: true,
+        enterpriseDomain,
         secrets: {},
         commandInfo,
       })
@@ -1464,6 +1527,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),
@@ -1495,6 +1560,8 @@ describe("pickEnvironment", () => {
           artifactsPath,
           branch: "main",
           username,
+          loggedIn: true,
+          enterpriseDomain,
           secrets: {},
           commandInfo,
         }),