+++ core/test/unit/src/config/template-contexts/project.ts
@@ -11,6 +11,7 @@ import stripAnsi = require("strip-ansi")
 import { ConfigContext } from "../../../../../src/config/template-contexts/base"
 import { ProjectConfigContext } from "../../../../../src/config/template-contexts/project"
 import { resolveTemplateString } from "../../../../../src/template-string/template-string"
+import { deline } from "../../../../../src/util/string"
 
 type TestValue = string | ConfigContext | TestValues | TestValueFunction
 type TestValueFunction = () => TestValue | Promise<TestValue>
@@ -19,6 +20,8 @@ interface TestValues {
 }
 
 describe("ProjectConfigContext", () => {
+  const enterpriseDomain = "https://garden.mydomain.com"
+
   it("should resolve local env variables", () => {
     process.env.TEST_VARIABLE = "value"
     const c = new ProjectConfigContext({
@@ -27,6 +30,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -43,6 +48,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -51,18 +58,86 @@ describe("ProjectConfigContext", () => {
     })
   })
 
-  it("should resolve secrets", () => {
+  it("should resolve when logged in", () => {
     const c = new ProjectConfigContext({
       projectName: "some-project",
       projectRoot: "/tmp",
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: { foo: "banana" },
       commandInfo: { name: "test", args: {}, opts: {} },
     })
     expect(c.resolve({ key: ["secrets", "foo"], nodePath: [], opts: {} })).to.eql({
       resolved: "banana",
+    })
+  })
+
+  context("errors thrown when a missing secret is referenced", () => {
+    it("should ask the user to log in if they're logged out", () => {
+      const c = new ProjectConfigContext({
+        projectName: "some-project",
+        projectRoot: "/tmp",
+        artifactsPath: "/tmp",
+        branch: "main",
+        username: "some-user",
+        loggedIn: false, // <-----
+        enterpriseDomain,
+        secrets: { foo: "banana" },
+        commandInfo: { name: "test", args: {}, opts: {} },
+      })
+
+      const { message } = c.resolve({ key: ["secrets", "bar"], nodePath: [], opts: {} })
+
+      expect(stripAnsi(message!)).to.match(/Please log in via the garden login command to use Garden with secrets/)
+    })
+
+    context("when logged in", () => {
+      it("should notify the user if an empty set of secrets was returned by the backend", () => {
+        const c = new ProjectConfigContext({
+          projectName: "some-project",
+          projectRoot: "/tmp",
+          artifactsPath: "/tmp",
+          branch: "main",
+          username: "some-user",
+          loggedIn: true,
+          enterpriseDomain,
+          secrets: {}, // <-----
+          commandInfo: { name: "test", args: {}, opts: {} },
+        })
+
+        const { message } = c.resolve({ key: ["secrets", "bar"], nodePath: [], opts: {} })
+
+        const errMsg = deline`
+          Looks like no secrets have been created for this project and/or environment in Garden Enterprise.
+          To create secrets, please visit ${enterpriseDomain} and navigate to the secrets section for this project.
+        `
+        expect(stripAnsi(message!)).to.match(new RegExp(errMsg))
+      })
+
+      it("if a non-empty set of secrets was returned by the backend, provide a helpful suggestion", () => {
+        const c = new ProjectConfigContext({
+          projectName: "some-project",
+          projectRoot: "/tmp",
+          artifactsPath: "/tmp",
+          branch: "main",
+          username: "some-user",
+          loggedIn: true,
+          enterpriseDomain,
+          secrets: { foo: "banana " }, // <-----
+          commandInfo: { name: "test", args: {}, opts: {} },
+        })
+
+        const { message } = c.resolve({ key: ["secrets", "bar"], nodePath: [], opts: {} })
+
+        const errMsg = deline`
+          Please make sure that all required secrets for this project exist in Garden Enterprise, and are accessible in this
+          environment.
+        `
+        expect(stripAnsi(message!)).to.match(new RegExp(errMsg))
+      })
     })
   })
 
@@ -73,6 +148,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -92,6 +169,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "some-user",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -107,6 +186,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "SomeUser",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -125,6 +206,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "SomeUser",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })
@@ -140,6 +223,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "SomeUser",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "deploy", args: {}, opts: { "hot-reload": ["my-service"] } },
     })
@@ -158,6 +243,8 @@ describe("ProjectConfigContext", () => {
       artifactsPath: "/tmp",
       branch: "main",
       username: "SomeUser",
+      loggedIn: true,
+      enterpriseDomain,
       secrets: {},
       commandInfo: { name: "test", args: {}, opts: {} },
     })